Refactor argument construction and improve error handling
Refactored gd::argument::arguments constructors to use common_construct_range for initializer lists and ranges, improving consistency and clarity. Replaced common_construct with common_construct_range in protected sections to avoid overload ambiguity. Fixed a potential signed/unsigned comparison in print_name_s and corrected an off-by-one error in string splitting logic. Enhanced error handling in FILES_Harvest_g for invalid file paths. Improved debug output and formatting in Application.cpp and CLIList.cpp. Added a new test case for argument construction in PLAY_Html.cpp. Performed minor code cleanups throughout.

+++ external/gd/gd_arguments.cpp
@@ -3379,7 +3379,7 @@ std::string arguments::print_s(const_pointer pPosition, uint32_t uPairType)
 
 void arguments::print_name_s( const_pointer pPosition, std::string& stringPrint )
 {
-   if( *pPosition == eType_ParameterName )
+   if( uint8_t(*pPosition) == eType_ParameterName )
    {
       // append name, name starts two bytes after position and size of name is in byte before name
       stringPrint.append( reinterpret_cast<const char*>( pPosition+ 2), *(pPosition + 1) );
